requirements.txt
Django
phonenumbers
django-localflavor
django-countries[pyuca]
django-phonenumber-field[phonenumberslite]
django-comuni-italiani
psycopg2
Pillow

